Rename InputDependencySourceMap -> OutputFileMap.DependencyView
Preset this API as a "view" of the output file map, rather than a separate data structure.
